Create Python Program to Merge Mails
For this program, we have written all the names in separate lines in the file "names.txt". The body is in the "body.txt" file.


We open both the files in reading mode and iterate over each name using a for loop. A new file with the name "[name].txt" is created, where name is the name of that person.

We use strip() method to clean up leading and trailing whitespaces (reading a line from the file also reads the newline '\n' character). Finally, we write the content of the mail into this file using the write() method.